+++ AoE2ScenarioRms/util/xs_util.py
@@ -8,11 +8,11 @@ class XsUtil:
         return 'true' if val else 'false'
 
     @staticmethod
-    def read(path: str) -> str:
+    def file(path: str) -> str:
         with (Path(__file__).parent.parent / 'xs' / path).open() as file:
             return file.read()
 
     @staticmethod
-    def format_name(name: str):
+    def constant(name: str):
         xs_name = ' '.join(filter(lambda e: e, name.split(' ')))
         return xs_name.upper().replace(' ', '_')